Refactor random number stuff out of benchmarks.
