+++ bigtable/benchmarks/benchmark.cc
@@ -14,6 +14,7 @@
 
 #include "bigtable/benchmarks/benchmark.h"
 #include "bigtable/benchmarks/random_mutation.h"
+#include "bigtable/client/table_admin.h"
 #include <future>
 #include <iomanip>
 #include <sstream>
@@ -120,7 +121,8 @@ BenchmarkResult Benchmark::PopulateTable() {
   return result;
 }
 
-std::string Benchmark::MakeRandomKey(DefaultPRNG& gen) const {
+std::string Benchmark::MakeRandomKey(
+    bigtable::testing::DefaultPRNG& gen) const {
   std::uniform_int_distribution<long> prng_user(0, setup_.table_size() - 1);
   return MakeKey(prng_user(gen));
 }
@@ -240,7 +242,7 @@ BenchmarkResult Benchmark::PopulateTableShard(bigtable::Table& table,
   BenchmarkResult result{};
   result.row_count = 0;
 
-  auto generator = MakeDefaultPRNG();
+  auto generator = bigtable::testing::MakeDefaultPRNG();
   int bulk_size = 0;
   bigtable::BulkMutation bulk;
