+++ bigtable/benchmarks/endurance_benchmark.cc
@@ -114,7 +114,7 @@ int main(int argc, char* argv[]) try {
 
 namespace {
 OperationResult RunOneApply(bigtable::Table& table, Benchmark const& benchmark,
-                            DefaultPRNG& generator) {
+                            bigtable::testing::DefaultPRNG& generator) {
   auto row_key = benchmark.MakeRandomKey(generator);
   bigtable::SingleRowMutation mutation(std::move(row_key));
   for (int field = 0; field != kNumFields; ++field) {
@@ -126,7 +126,7 @@ OperationResult RunOneApply(bigtable::Table& table, Benchmark const& benchmark,
 
 OperationResult RunOneReadRow(bigtable::Table& table,
                               Benchmark const& benchmark,
-                              DefaultPRNG& generator) {
+                              bigtable::testing::DefaultPRNG& generator) {
   auto row_key = benchmark.MakeRandomKey(generator);
   auto op = [&table, &row_key]() {
     auto row = table.ReadRow(
@@ -146,7 +146,7 @@ long RunBenchmark(bigtable::benchmarks::Benchmark& benchmark,
   auto data_client = benchmark.MakeDataClient();
   bigtable::Table table(std::move(data_client), table_id);
 
-  auto generator = MakeDefaultPRNG();
+  auto generator = bigtable::testing::MakeDefaultPRNG();
 
   auto start = std::chrono::steady_clock::now();
   auto end = start + test_duration;
