+++ bigtable/benchmarks/random_mutation.cc
@@ -17,16 +17,17 @@
 
 namespace bigtable {
 namespace benchmarks {
-bigtable::Mutation MakeRandomMutation(DefaultPRNG& gen, int f) {
+bigtable::Mutation MakeRandomMutation(bigtable::testing::DefaultPRNG& gen,
+                                      int f) {
   std::string field = "field" + std::to_string(f);
   return bigtable::SetCell(kColumnFamily, std::move(field),
                            std::chrono::milliseconds(0), MakeRandomValue(gen));
 }
 
-std::string MakeRandomValue(DefaultPRNG& generator) {
+std::string MakeRandomValue(bigtable::testing::DefaultPRNG& generator) {
   static std::string const letters(
       "ABCDEFGHIJLKMNOPQRSTUVWXYZabcdefghijlkmnopqrstuvwxyz0123456789-/_");
-  return Sample(generator, kFieldSize, letters);
+  return bigtable::testing::Sample(generator, kFieldSize, letters);
 }
 }  // namespace benchmarks
 }  // namespace bigtable
