+++ bigtable/benchmarks/scan_throughput_benchmark.cc
@@ -124,7 +124,7 @@ BenchmarkResult RunBenchmark(bigtable::benchmarks::Benchmark const& benchmark,
 
   bigtable::Table table(std::move(data_client), table_id);
 
-  auto generator = MakeDefaultPRNG();
+  auto generator = bigtable::testing::MakeDefaultPRNG();
   std::uniform_int_distribution<long> prng(0, table_size - scan_size - 1);
 
   auto test_start = std::chrono::steady_clock::now();
