+++ bigtable/client/testing/random.cc
@@ -1,4 +1,4 @@
-// Copyright 2017 Google Inc.
+// Copyright 2018 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,11 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "bigtable/benchmarks/random.h"
-#include "bigtable/benchmarks/constants.h"
+#include "bigtable/client/testing/random.h"
 
 namespace bigtable {
-namespace benchmarks {
+namespace testing {
 std::string Sample(DefaultPRNG& gen, int n, std::string const& population) {
   std::uniform_int_distribution<std::size_t> rd(0, population.size() - 1);
 
@@ -26,5 +25,5 @@ std::string Sample(DefaultPRNG& gen, int n, std::string const& population) {
   return result;
 }
 
-}  // namespace benchmarks
+}  // namespace testing
 }  // namespace bigtable
