+++ bigtable/client/testing/random.h
@@ -12,16 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef GOOGLE_CLOUD_CPP_BIGTABLE_BENCHMARKS_RANDOM_H_
-#define GOOGLE_CLOUD_CPP_BIGTABLE_BENCHMARKS_RANDOM_H_
+#ifndef GOOGLE_CLOUD_CPP_BIGTABLE_CLIENT_TESTING_RANDOM_H_
+#define GOOGLE_CLOUD_CPP_BIGTABLE_CLIENT_TESTING_RANDOM_H_
 
-#include "bigtable/client/table.h"
-#include "bigtable/client/table_admin.h"
 #include <algorithm>
 #include <random>
 
 namespace bigtable {
-namespace benchmarks {
+namespace testing {
 // While std::mt19937_64 is not the best PRNG ever, it is fairly good for
 // most purposes.  Please read:
 //    http://www.pcg-random.org/
@@ -78,7 +76,7 @@ inline DefaultPRNG MakeDefaultPRNG() { return MakePRNG<DefaultPRNG>(); }
  */
 std::string Sample(DefaultPRNG& gen, int n, std::string const& population);
 
-}  // namespace benchmarks
+}  // namespace testing
 }  // namespace bigtable
 
-#endif  // GOOGLE_CLOUD_CPP_BIGTABLE_BENCHMARKS_RANDOM_H_
+#endif  // GOOGLE_CLOUD_CPP_BIGTABLE_CLIENT_TESTING_RANDOM_H_
